A field is a class of the type of its value, and FixTag says which
The 912 per-tag field classes are gone. A field is one of eleven classes of
the type of its value -- FixField.Text, Character, Boolean, Integer, Decimal,
Timestamp, Time, Date, MonthYear, Multiple and Data, each Typed<T> -- or an
Invalid, and its Tag says which field it is: FixField.Decimal { Tag:
FixTag.OrderQty }.

- FixTag is an enum of the 912 standard tags. FixField.Tag, FixFinding.Tag,
  FixParseError.Tag and LengthDataPairs are typed by it; a tag the standard
  does not name is its number, (FixTag)25005.
- FixCustomField, FixCustomField<T>, the FixFieldFactory delegate and the
  field factory Load compiled through the expression language are removed.
- The context holds one table, a byte a tag: the type of its value and its
  half of a pair. The standard's is the default; Load types the fields a
  dictionary adds and never retypes a standard tag; a pair's tags nothing
  types are read as an Integer and a Data. A tag nothing defines is Invalid.
  FixFieldBuilder is a switch of twelve arms over the type.
- generate.py names the tags from the repository (QuickFIX's two names where
  they differ), types them, and writes FixTag.cs, the standard's table
  (FixFieldBuilder.Standard.cs) and the message and component properties
  typed by value class. The field checks are slots per tag, typed by value
  class, and Load writes its texts to those signatures.
- The oracle's grammar, the hand parsers, the stands and the package smoke
  follow; tests match on the value class and the tag. README, SKILL and the
  release notes describe the new model.

DotGram.Finance.Tests 4080 and DotGram.Finance.Fix44.Tests 613 pass; the
solution builds in Release without warnings.

+++ benchmarks/DotGram.Benchmarks/Stand.cs
@@ -503,7 +503,7 @@ static partial class Stand
 			var count = 0;
 
 			foreach (var field in fields)
-				count += field.Tag;
+				count += (int)field.Tag;
 
 			return count;
 		}
@@ -1625,7 +1625,7 @@ static partial class Stand
 			var count = 0;
 
 			foreach (var field in fields)
-				count += field.Tag;
+				count += (int)field.Tag;
 
 			return count;
 		}
